+++ litebird_sim/__init__.py
@@ -4,6 +4,7 @@ import numba
 
 from litebird_sim.mapmaking import (
     make_binned_map,
+    make_brahmap_gls_map,
     check_valid_splits,
     BinnerResult,
     make_destriped_map,
@@ -34,6 +35,7 @@ from .coordinates import (
     ECL_TO_GAL_ROT_MATRIX,
     CoordinateSystem,
     coord_sys_to_healpix_string,
+    rotate_coordinates_e2g,
 )
 from .detectors import (
     DetectorInfo,
@@ -51,14 +53,26 @@ from .gaindrifts import (
     apply_gaindrift_to_tod,
     apply_gaindrift_to_observations,
 )
+from .grasp2alm import (
+    BeamCut,
+    BeamGrid,
+    BeamHealpixMap,
+    BeamStokesPolar,
+    ticra_cut_to_alm,
+    ticra_grid_to_alm,
+)
 from .healpix import (
+    UNSEEN_PIXEL_VALUE,
     nside_to_npix,
     npix_to_nside,
+    nside_to_pixel_solid_angle_sterad,
+    nside_to_resolution_rad,
     is_npix_ok,
     map_type,
     get_pixel_format,
     write_healpix_map_to_hdu,
     write_healpix_map_to_file,
+    num_of_alms,
 )
 from .hwp import (
     HWP,
@@ -81,7 +95,7 @@ from .io import (
     read_list_of_observations,
 )
 from .madam import save_simulation_for_madam
-from .mbs.mbs import Mbs, MbsParameters, MbsSavedMapInfo
+from .mbs.mbs import FG_MODELS, Mbs, MbsParameters, MbsSavedMapInfo
 from .mpi import MPI_COMM_WORLD, MPI_ENABLED, MPI_CONFIGURATION, MPI_COMM_GRID
 from .mueller_convolver import MuellerConvolver
 from .noise import (
@@ -199,13 +213,17 @@ __all__ = [
     # bandpasses.py
     "BandPassInfo",
     # healpix.py
+    "UNSEEN_PIXEL_VALUE",
     "nside_to_npix",
     "npix_to_nside",
+    "nside_to_pixel_solid_angle_sterad",
+    "nside_to_resolution_rad",
     "is_npix_ok",
     "map_type",
     "get_pixel_format",
     "write_healpix_map_to_hdu",
     "write_healpix_map_to_file",
+    "num_of_alms",
     # distribute.py
     "distribute_evenly",
     "distribute_optimally",
@@ -221,6 +239,13 @@ __all__ = [
     "FreqChannelInfo",
     "InstrumentInfo",
     "detector_list_from_parameters",
+    # grasp2alm.py
+    "BeamCut",
+    "BeamGrid",
+    "BeamHealpixMap",
+    "BeamStokesPolar",
+    "ticra_cut_to_alm",
+    "ticra_grid_to_alm",
     # hwp.py
     "HWP",
     "IdealHWP",
@@ -230,6 +255,7 @@ __all__ = [
     # madam.py
     "save_simulation_for_madam",
     # mbs.py
+    "FG_MODELS",
     "Mbs",
     "MbsParameters",
     "MbsSavedMapInfo",
@@ -294,6 +320,7 @@ __all__ = [
     "apply_hwp_to_obs",
     # mapmaking
     "make_binned_map",
+    "make_brahmap_gls_map",
     "check_valid_splits",
     "BinnerResult",
     "make_destriped_map",
@@ -324,6 +351,7 @@ __all__ = [
     "ECL_TO_GAL_ROT_MATRIX",
     "CoordinateSystem",
     "coord_sys_to_healpix_string",
+    "rotate_coordinates_e2g",
     # spacecraft.py
     "compute_l2_pos_and_vel",
     "compute_lissajous_pos_and_vel",