+++ litebird_sim/hwp.py
@@ -135,11 +135,16 @@ class HWP:
 def _get_ideal_hwp_angle(
     output_buffer, start_time_s, delta_time_s, start_angle_rad, ang_speed_radpsec
 ):
-    for sample_idx in range(output_buffer.size):
-        angle = (
-            start_angle_rad
-            + (start_time_s + delta_time_s * sample_idx) * ang_speed_radpsec
-        ) % (2 * np.pi)
+    complex = np.exp(1j * start_angle_rad) * np.exp(
+        1j * start_time_s * ang_speed_radpsec
+    )
+    delta = np.exp(1j * delta_time_s * ang_speed_radpsec)
+    output_buffer[0] = np.angle(complex)
+
+    for sample_idx in range(1, output_buffer.size):
+        complex *= delta
+        complex /= np.sqrt(complex.real**2 + complex.imag**2)
+        angle = np.angle(complex)
 
         output_buffer[sample_idx] = angle
 
@@ -212,21 +217,6 @@ class IdealHWP(HWP):
             delta_time_s=delta_time_s,
             start_angle_rad=self.start_angle_rad,
             ang_speed_radpsec=self.ang_speed_radpsec,
-        )
-
-    def apply_hwp_to_pointings(
-        self,
-        start_time_s: float,
-        delta_time_s: float,  # This will be 1/(19 Hz) in most cases
-        bore2ecl_quaternions_inout: npt.NDArray,  # Boresight→Ecliptic quaternions at 19 Hz
-        hwp_angle_out: npt.NDArray,
-    ) -> None:
-        # We do not touch `bore2ecl_quaternions_input`, as an ideal HWP does not
-        # alter the (θ, φ) direction of the boresight nor the orientation ψ
-        self.get_hwp_angle(
-            output_buffer=hwp_angle_out,
-            start_time_s=start_time_s,
-            delta_time_s=delta_time_s,
         )
 
     def write_to_hdf5(