+++ CHANGELOG.md
@@ -1,5 +1,31 @@
 # HEAD
 
+-   Fix issue [#457](https://github.com/litebird/litebird_sim/issues/457)
+
+-   Improve the documentation [#450](https://github.com/litebird/litebird_sim/pull/450)
+
+-   New method `Imo.open_data_file` [#449](https://github.com/litebird/litebird_sim/pull/449)
+
+-   Improved `SphericalHarmonics` class, algebra and I/O implemented [#448](https://github.com/litebird/litebird_sim/pull/448)
+
+-   Import Yusuke Takase's grasp2alm code [#447](https://github.com/litebird/litebird_sim/pull/447)
+
+-   Change hwp_angle calculation to the complex domain [#433](https://github.com/litebird/litebird_sim/pull/433)
+
+# Version 0.15.3
+
+-   Fix the computation of HWP angles [#444](https://github.com/litebird/litebird_sim/pull/444)
+
+-   Add low-level interface to `BrahMap` [#440](https://github.com/litebird/litebird_sim/pull/440)
+
+-   Set Mueller matrix phases in Hwp_sys module as class attributes, instead of being hardcoded [#442](https://github.com/litebird/litebird_sim/pull/442)
+
+# Version 0.15.2
+
+-   Add more tutorials [#443](https://github.com/litebird/litebird_sim/pull/443)
+
+# Version 0.15.1
+
 -   Fixed s4,s5,d9,d10 and added s7 in Mbs [#437](https://github.com/litebird/litebird_sim/pull/437)
 
 -   Make sure that the PTEP IMo, the Madam templates, and the static files required to create the HTML reports are installed properly by `pip` and enable Binder/Google Colab [#436](https://github.com/litebird/litebird_sim/pull/436)